Fix state_dict loading and 5 failing CI tests
Add _assign_state_dict helper in model.py that loads state dicts via
direct setattr instead of nn.Module.load_state_dict. This is necessary
because PyTorch rejects shape mismatches even with assign=True -- and
TurboQuantLinear registers empty buffers at init that get populated from
checkpoint data with different shapes.

The helper distinguishes parameters (wrapped in nn.Parameter) from
buffers (plain tensors) to satisfy PyTorch's module attribute type
checking.

Test fixes:
- test_state_dict_serialization: use _assign_state_dict
- test_state_dict_serialization_remainder_only: same
- test_save_load_roundtrip_produces_same_output: build TurboQuantLinear
  shells manually instead of re-quantizing (which produced different
  weights from different random init)
- test_batch_compression_correctness: use allclose instead of exact equal
  for float32-to-float16 window comparison
- test_invalid_compute_dtype: test with "int8" since float16 is now a
  valid compute_dtype

Verified: 78 passed, 0 failed, ruff lint clean.

+++ tests/test_quantizer.py
@@ -138,7 +138,7 @@ class TestTurboQuantConfig:
 
     def test_invalid_compute_dtype(self):
         with pytest.raises(ValueError):
-            TurboQuantConfig(compute_dtype="float16")
+            TurboQuantConfig(compute_dtype="int8")
 
     def test_serialization(self, tmp_path):
         config = TurboQuantConfig(bit_width=3, group_size=64, rotation_seed=123)
